Rename experimental FSDP runtime types
